+++ Simulator/src/main/java/ui/modeleditor/elements/ModelElementSubAnimationDashboardWindow.java
@@ -80,7 +80,7 @@ public class ModelElementSubAnimationDashboardWindow extends JFrame implements R
 	 * Gewählter Delay-Wert
 	 * @see AnimationPanel#getDelayIntern()
 	 */
-	private int delayInt;
+	private double delayInt;
 	/**
 	 * Simulationsdatenobjekt
 	 * @see #updateViewer(SimulationData)
@@ -176,7 +176,7 @@ public class ModelElementSubAnimationDashboardWindow extends JFrame implements R
 			if (surfaceAnimator.testBreakPoints(simData,client)) {
 				mainAnimationPanel.playPause();
 				surfaceAnimator.updateSurfaceAnimationDisplayElements(simData,true,false);
-				if (!moveByTransport) surfaceAnimator.process(simData,client,FastMath.min(20,delayInt/4));
+				if (!moveByTransport) surfaceAnimator.process(simData,client,FastMath.min(20,delayInt/4.0));
 				surfacePanel.repaint();
 			}
 		}
@@ -193,7 +193,7 @@ public class ModelElementSubAnimationDashboardWindow extends JFrame implements R
 		this.simData=simData;
 
 		/* keine normale Verzögerung hier, da diese schon auf der Hauptebene erfolgt */
-		if (!moveByTransport) surfaceAnimator.process(simData,client,FastMath.min(20,delayInt/4));
+		if (!moveByTransport) surfaceAnimator.process(simData,client,FastMath.min(20,delayInt/4.0));
 		return true;
 	}
 
@@ -213,7 +213,7 @@ public class ModelElementSubAnimationDashboardWindow extends JFrame implements R
 		this.simData=simData;
 
 		/* keine normale Verzögerung hier, da diese schon auf der Hauptebene erfolgt */
-		surfaceAnimator.process(simData,transporter,FastMath.min(20,delayInt/4));
+		surfaceAnimator.process(simData,transporter,FastMath.min(20,delayInt/4.0));
 		return true;
 	}
